Functioning movement and sinking light bug fixed!
