+++ AI Game Jam/Assets/Scripts/SecurityLightMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6f9533d7b8177b647a802a8ea842ec24
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
